Move Clipboard to an Extension, cut out "REBREQ"
This migrates the clipboard port to be loaded as an extension, and the
code for the extension is written directly to the OS API.  This
cuts out the "Device" layer, by saying that the only relevant layer
of abstraction is the PORT! abstraction itself.

Should another OS wish to have clipboard abilities, it would #ifdef in
the code or otherwise implement the PORT! in its own way, without
having to go through the device request model.

Message pumps may still be required, but that could be done more
narrowly through EVENT! features.
