+++ make/tools/file-base.r
@@ -92,7 +92,6 @@ core: [
     n-textcodecs.c ; !!! should be moved to extensions
 
     ; (P)orts
-    p-clipboard.c
     p-console.c
     p-dir.c
     p-dns.c
@@ -197,7 +196,6 @@ os-windows: [
     windows/dev-stdio.c
     windows/dev-file.c
     windows/dev-event.c
-    windows/dev-clipboard.c
     windows/dev-serial.c
 ]
